server/requirements.txt
SpeechRecognition
--extra-index-url https://download.pytorch.org/whl/cu116
torch
git+https://github.com/openai/whisper.git
pydub
flask
cryptography
python-dotenv
prometheus_client